requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
python-dotenv
tiktoken
faiss-cpu
chromadb
bs4